Made the database persistent
Until now, each docker-compose down will loose all data (jobs, devices, etc...).
By creating a docker volume, the database will now be persistent other restart.
Due to a risk of the lava package reseting the password, it is always reseted at start.
This patch add also another volume for job output which are stored in the filesystem and not in database.