Split dockerfile in two images
The process of building LAVA is hard to maintain.
But since we need to be able to go back in time
(or just stay longer with a specific version)
we need a way to keep a LAVA version.
For achieving this, we build a minimal image with just LAVA
and tag it with the version of LAVA inside.