lava-slave/scripts: fix master key name
The ZMQ auth scripts are copying the master key to the slave config
with a hard-coded file name: master.key.  Yet the lava-slave config
file was looking for $LAVA_MASTER.key.

This caused the slave to fail to find the key and abort startup.

Fix by using the same hard-coded filename in the lava-slave config.