lavalab-gen.py: replace slave_name by worker_name
In the board loop, the slave name is stored in slave_name, but some code
after use worker_name.
This works when only one slave exists, but fail when it exists multiple
slave.
This patch replace slave_name by worker_name for being consistent with
the naming used in the rest of code.