lava-slave: handle the case where there are no entrypoints to copy
If there are no entrypoint to copy, the COPY docker action will fail.
For fixing this issue, simply copy all files instead of just "*sh".
Since the entrypoint.d directory is not empty (got an empty file) it
will always works.
The LAVA script run only *sh files so we are still safe.