Make docker bind the TFTP server to the correct interface
With UDP, docker sometimes can't route UDP packets to the correct
address because the source address gets mangled [1].  This happens
indeterministically because by default the kernel uses some heuristics
to decide which interface to use.  But in some rare circumstances, those
can fail and a wrong source address is used.  This can be worked around
by explicitly binding to a concrete interface instead of 0.0.0.0.

[1]: https://github.com/moby/libnetwork/issues/1729
Signed-off-by: Harald Seiler <hws@denx.de>