[Fix] Ensure proper filename of public master key
Currently, the public master key is references as $LAVA_MASTER.key in
e.g. scripts/setup.sh. If the key is named master.key it is copied to
$LAVA_MASTER.key by scripts/setup.sh. However, no nameing convention for the
public master key provided to lavalab-gen.py is documented, so the worker simply
will not start if the key is stored as e.g. master-public.key. Running into such
error scenarios could simply be prevented by copying the public master key to an
appropriatly named file in lavalab-gen.py. This is done here.