Permit to set worker token in boards.yaml
LAVA 2020.09 dropped ZMQ and use now a token to authenticate worker
against master.

lava-docker already handle this by getting token via lavacli.
But we need to support to set token in boards.yaml, this patch permits
this.

Furthermore, the token was given to worker via --token, but this is bad
since token can be found by anyone running ps.
A better secure way is to use --token-file.