Test upgrade of 2023.06
LAVA from 2023.01 to 2023.05 miss a migration script which is generated by
lava-docker but with a "automatic" name.
LAVA 2023.06 has the correct migration script but you cannot have both
at the same time.
Add a detection of this situation to rollback automatic migration.

In the same time replace all migration action via the postinst.py which
do the correct action.

Signed-off-by: Corentin LABBE <clabbe@baylibre.com>