Use lavacli to set the dispatcher IP
Currently when configuring the dispatcher IP for a worker we use a
custom script setdispatcherip.py to dynamically generate a config and do
an XMLRPC call to set a configuration for the worker.  This is a bit
limiting if we want to set any other configuration for the worker,
instead let's write a configuration file into the worker container at
build time and then use lavacli (which we already rely on) to apply
that during startup.

Signed-off-by: Mark Brown <broonie@kernel.org>