Support KissCache style proxies
LAVA supports an alternative cache style where it just passes the URL to
be fetched to a web service and accepts whatever the web service
provides as valid, allowing caching of objects served over HTTPS without
complicated certificate management.  This is configured in the worker
config with the 'http_url_format_string' parameter, support doing so via
lava-docker.

Signed-off-by: Mark Brown <broonie@kernel.org>